Button component moved to the commom folder
